Feature #13830: improving survey result pages.
* modifying the getPercent signature of Answer class in order to return a BigDecimal instance
* adding to SpVue API the ability to resolve and render dynamically a registered component
* improving the Silverpeas's chart JS plugin

+++ core-services/questioncontainer/src/main/java/org/silverpeas/core/questioncontainer/answer/model/Answer.java
@@ -160,9 +160,9 @@ public class Answer implements Serializable {
     return this.questionLink;
   }
 
-  public double getPercent(int nbParticipants) {
+  public BigDecimal getPercent(int nbParticipants) {
     return new BigDecimal(getNbVoters()).multiply(new BigDecimal(100))
-        .divide(new BigDecimal(nbParticipants), 2, RoundingMode.HALF_DOWN).doubleValue();
+        .divide(new BigDecimal(nbParticipants), 2, RoundingMode.HALF_DOWN);
   }
 
   @Override
